Sort releases for CSV report
